pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "norman-mcp-server"
version = "0.1.4"
description = "A Model Context Protocol (MCP) server for Norman Finance API"
readme = "README.md"
authors = [
    {name = "Norman Finance", email = "stan@norman.finance"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.3.0",
    "mcp[cli]>=1.7.0",
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=22.3.0",
    "isort>=5.10.1",
    "flake8>=4.0.0",
    "mypy>=0.942",
]
sse = [
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/norman-finance/norman-mcp-server"
Issues = "https://github.com/norman-finance/norman-mcp-server/issues"

[project.scripts]
norman-mcp = "norman_mcp.cli:main"

[tool.setuptools]
packages = ["norman_mcp"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
